Initial version of the computer algebra system Leibniz.
This is a quick sketch of a computer algebra system interface that I created
after some discussion with Pavanakumar Mohanamuraly and Robert Smith.

It is mostly intended to start a discussion, and give some ideas about how a
CLOSsy math interface could look like.  None of the design decisions is set in
stone.